Don't overwrite fields from the lumberjack kv set.
This should fix a bug where a user uses `codec => json` and has
a field named "host" in this json. The bug is that logstash-forwarder
will send a "host" field along with every event, and the lumberjack
input would previously ignore any previous value for the "host" field.
This is also fixed for "file" and "offset" fields.

Fixes #24